Proper obsolete of AccountRoutingSettings
