+++ src/Tests/ApprovalFiles/APIApprovals.ApproveAzureStorageQueueTransport.approved.txt
@@ -10,22 +10,13 @@ namespace NServiceBus
         public string ConnectionString { get; }
         public System.Collections.Generic.HashSet<string> RegisteredEndpoints { get; }
     }
-    [System.Obsolete("Account aliases have been deprecated. Use the AzureStorageQueueTransport construc" +
-        "tor that accepts fully customized Azure service clients. Will be treated as an e" +
-        "rror from version 11.0.0. Will be removed in version 12.0.0.", false)]
     public class AccountRoutingSettings
     {
-        [System.Obsolete("Account aliases have been deprecated. Use the AzureStorageQueueTransport construc" +
-            "tor that accepts fully customized Azure service clients. Will be treated as an e" +
-            "rror from version 11.0.0. Will be removed in version 12.0.0.", false)]
         public string DefaultAccountAlias { get; set; }
-        [System.Obsolete("Account aliases have been deprecated. Use the AzureStorageQueueTransport construc" +
-            "tor that accepts fully customized Azure service clients. Will be treated as an e" +
-            "rror from version 11.0.0. Will be removed in version 12.0.0.", false)]
         public NServiceBus.AccountInfo AddAccount(string alias, Azure.Storage.Queues.QueueServiceClient connectionClient) { }
-        [System.Obsolete("Account aliases have been deprecated. Use the AzureStorageQueueTransport construc" +
-            "tor that accepts fully customized Azure service clients. Will be treated as an e" +
-            "rror from version 11.0.0. Will be removed in version 12.0.0.", false)]
+        [System.Obsolete("Account aliases using connection strings have been deprecated. Use the AddCount o" +
+            "verload that accepts a QueueServiceClient instance. Will be treated as an error " +
+            "from version 11.0.0. Will be removed in version 12.0.0.", false)]
         public NServiceBus.AccountInfo AddAccount(string alias, string connectionString) { }
     }
     public class AzureStorageQueueTransport : NServiceBus.Transport.TransportDefinition, NServiceBus.Routing.IMessageDrivenSubscriptionTransport
