+++ src/Transport/Config/Extensions/AccountRoutingSettings.cs
@@ -1,0 +1,53 @@
+﻿namespace NServiceBus
+{
+    using System;
+    using System.Collections.Generic;
+    using global::Azure.Storage.Queues;
+
+    /// <summary>
+    /// Provides methods to define routing between Azure Storage accounts and map them to a logical alias instead of using bare
+    /// connection strings.
+    /// </summary>
+    public partial class AccountRoutingSettings
+    {
+        internal AccountRoutingSettings()
+        {
+
+        }
+
+        /// <summary>
+        /// Get or set default account alias.
+        /// </summary>
+        public string DefaultAccountAlias
+        {
+            get => _defaultAccountAlias;
+            set
+            {
+                if (string.IsNullOrWhiteSpace(value))
+                {
+                    throw new ArgumentException("Should not be null or white space", nameof(DefaultAccountAlias));
+                }
+                _defaultAccountAlias = value;
+            }
+        }
+
+        /// <summary>
+        /// Adds the mapping between the <paramref alias="alias" /> and its <paramref alias="QueueServiceClient" />.
+        /// </summary>
+        public AccountInfo AddAccount(string alias, QueueServiceClient connectionClient)
+        {
+            if (mappings.TryGetValue(alias, out var accountInfo))
+            {
+                return accountInfo;
+            }
+
+            accountInfo = new AccountInfo(alias, connectionClient);
+            mappings.Add(alias, accountInfo);
+
+            return accountInfo;
+        }
+
+        internal Dictionary<string, AccountInfo> mappings = new Dictionary<string, AccountInfo>();
+        string _defaultAccountAlias = string.Empty;
+    }
+}
